+++ model/src/main/java/edu/rsatu/garage/entities/Model.java
@@ -1,23 +1,23 @@
 package edu.rsatu.garage.entities;
 
 public class Model {
-    private Integer id;
+    private Long id;
     private String name;
 
     public Model(String name) {
         this.name = name;
     }
 
-    public Model(Integer id, String name) {
+    public Model(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
